Simplify inline positioning
This reverts some of the recently introduced logic to handle
inline-frame positioning, and instead integrates the positioning into
text layout. Decorations at the end of an inline element are not
considered for now. I think that can better be solved together
with properly handling decorations on inline elements themselves
(instead of copying them to their text-node children).