Add basic layout tests for resolving image dimensions
As an alternative to the existing FrameReflower tests.

Differences in expected sizes for most of the tests is due to the
FrameReflower tests not mocking DPI for the Style class, so calculations
via `length_in_pt` use the default 96 DPI, while `auto` image
dimensions are resolved with the mocked 75 DPI.

The layout tests use the default DPI instead to avoid failing tests due
to static caches (in `length_in_pt`) when running the entire test suite
at once. This could be solved with the `@runTestsInSeparateProcesses`/
`@runInSeparateProcess` annotations, but that makes testing noticeably
slower.